move responsibility to NodeDecorators and use glue in MetadataNodeVisitor only
