+++ packages/NodeTypeResolver/src/Metadata/NamespaceNodeDecorator.php
@@ -1,15 +1,15 @@
 <?php declare(strict_types=1);
 
-namespace Rector\NodeTypeResolver\NodeVisitor;
+namespace Rector\NodeTypeResolver\Metadata;
 
 use PhpParser\Node;
 use PhpParser\Node\Stmt\Namespace_;
 use PhpParser\Node\Stmt\Use_;
-use PhpParser\NodeVisitorAbstract;
 use Rector\NodeTraverserQueue\BetterNodeFinder;
+use Rector\NodeTypeResolver\Contract\Metadata\NodeDecoratorInterface;
 use Rector\NodeTypeResolver\Node\MetadataAttribute;
 
-final class NamespaceResolverNodeVisitor extends NodeVisitorAbstract
+final class NamespaceNodeDecorator implements NodeDecoratorInterface
 {
     /**
      * @var string|null
@@ -36,17 +36,14 @@ final class NamespaceResolverNodeVisitor extends NodeVisitorAbstract
         $this->betterNodeFinder = $betterNodeFinder;
     }
 
-    /**
-     * @param Node[] $nodes
-     */
-    public function beforeTraverse(array $nodes): void
+    public function reset(): void
     {
         $this->namespaceName = null;
         $this->namespaceNode = null;
         $this->useNodes = [];
     }
 
-    public function enterNode(Node $node): void
+    public function decorateNode(Node $node): void
     {
         if ($node instanceof Namespace_) {
             $this->namespaceName = $node->name ? $node->name->toString() : null;
