+++ src/config/services.yml
@@ -21,5 +21,4 @@ services:
             - ['addNodeVisitor', ['@PhpParser\NodeVisitor\NameResolver']]
             - ['addNodeVisitor', ['@PhpParser\NodeVisitor\CloningVisitor']]
             - ['addNodeVisitor', ['@Rector\PhpParser\NodeVisitor\ParentAndNextNodeAddingNodeVisitor']]
-            - ['addNodeVisitor', ['@Rector\NodeTypeResolver\NodeVisitor\ClassAndMethodResolverNodeVisitor']]
-            - ['addNodeVisitor', ['@Rector\NodeTypeResolver\NodeVisitor\NamespaceResolverNodeVisitor']]
+            - ['addNodeVisitor', ['@Rector\NodeTypeResolver\NodeVisitor\MetadataNodeVisitor']]
